% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{create.removal.Observer}
\alias{create.removal.Observer}
\title{Create data object for removal analysis with distance}
\usage{
create.removal.Observer(transect.dat, obs.dat)
}
\arguments{
\item{transect.dat}{Data table of each transect.}

\item{obs.dat}{Data table observations in each transect.}
}
\value{
Data table with counts summarized on each transect
}
\description{
A helper function to format the data appropriately for a ddf analysis using the R package `MRDS`.
}
\examples{
 
create.removal.Observer(transect.dat=transect.dat, obs.dat=distance.dat) 
}
