% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFuncs.R
\name{create.removal}
\alias{create.removal}
\title{Create data object for removal analysis without distance}
\usage{
create.removal(double.dat, transect.dat, pool.transects = TRUE)
}
\arguments{
\item{transect.dat}{Data table of each transect.}

\item{pool.transects}{Analyze transects seperately (FALSE) or together (TRUE)}

\item{obs.dat}{ata table observations in each transect.}
}
\value{
A list object with counts summarized on each transect or over all transects
}
\description{
A helper function to format the data appropriately for a removal analysis the R package `FSA`.
}
\examples{
create.removal(transect.dat, obs.dat, pool.transects=TRUE)
}
