% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DistanceEncounter}
\alias{DistanceEncounter}
\title{Encounters of zebra mussels in Lake Burgen, MN}
\format{\describe{
    \item{Observer name}{The name of the observer that made the detection}
    \item{Transect number}{The transect the detection was made on}
    \item{Distance}{The distance of the detection event from the transect line.}
    \item{size}{The number of individual mussels in the detection event.}
}}
\source{
\href{https://github.com/troutinthemilk/Zeebs-Season2/blob/master/Tutorial/EncountersDistanceTutorial.xlsx}{github}, based on data from \href{https://conservancy.umn.edu/handle/11299/201572}{Data Repository for University of Minnesota}
}
\usage{
DistanceEncounter
}
\description{
Number of zebra mussel encounters made by two observers in a double-observer distance design.
}
\examples{
data(DistanceEncounter)
}
\keyword{datasets}
