% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DoubleEncounter}
\alias{DoubleEncounter}
\title{Encounters of zebra mussels in Lake Burgen, MN}
\format{A data frame with 53940 rows and 10 variables:}
\source{
\href{https://github.com/troutinthemilk/Zeebs-Season2/blob/master/Tutorial/EncountersDoubleTutorial.xlsx}{github}, based on data from \href{https://conservancy.umn.edu/handle/11299/201572}{Data Repository for University of Minnesota}
}
\usage{
DoubleEncounter
}
\description{
Number of zebra mussel encounters made by two observers in a double-observer removal design.
}
\examples{
data(DoubleEncounter)
}
\keyword{datasets}
