% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DoubleTransect}
\alias{DoubleTransect}
\title{Transect information for zebra mussels in Lake Burgen, MN}
\format{\describe{
    \item{primary}{The name of the primary observer on this transect}
    \item{secondary}{The name of the secondary observer on this transect}
    \item{length}{The length of the current transect}
    \item{Transect number}{The current transect}
}}
\source{
\href{https://github.com/troutinthemilk/Zeebs-Season2/blob/master/Tutorial/EncountersDoubleTransect.xlsx}{github}, based on data from \href{https://conservancy.umn.edu/handle/11299/201572}{Data Repository for University of Minnesota}
}
\usage{
DoubleTransect
}
\description{
Transects surveyed by two observers in a double-observer removal design.
}
\examples{
data(DoubleTransect)
}
\keyword{datasets}
